all last changes saved

+++ SimGen2/make.sh
@@ -1,5 +1,7 @@
-rm sims.xml
-ant jar
+if [ -f sims.xml ]; then rm sims.xml; fi
+echo -ne '\e[1;30m'
+ant jar -q
 java -jar dist/SimGen3.jar
-cat sims.xml
+echo -ne '\e[0m';
+cat sims.xml | less
 echo
